modify: files to handle tauri v2
